feat: create MoustRecent component and move Footer to Areas directory; add styles for Footer and MainHub
