+++ components/Areas/MoustRecent/index.tsx
@@ -1,0 +1,12 @@
+import Card from '@/components/Areas/Card'
+
+const MoustRecent = () => {
+  return (
+    <div className="card-area">
+      <Card></Card>
+      <Card></Card>
+    </div>
+  )
+}
+
+export default MoustRecent
