--- containers/MainHub/index.tsx
@@ -1,126 +1,0 @@
-import './styles.css'
-const MainHub = () => {
-  return (
-    <div className="main-hub">
-      <div className="update-area">
-        <h2>Update Area</h2>
-        <h1>Version 1.0</h1>
-        <p className="update-time">January 1, 2024 by USER</p>
-        <div className="update-content">
-          <p>
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas ut
-            mattis ligula. Maecenas congue, ipsum nec congue scelerisque, risus
-            arcu placerat ligula, at fringilla massa justo ut erat. Mauris sed
-            tellus est. Aenean fringilla, lectus vitae vehicula sollicitudin,
-            mauris ex tempus dui, in consectetur nisl quam ac felis. Donec eu
-            laoreet velit, sed fermentum lectus. Pellentesque habitant morbi
-            tristique senectus et netus et malesuada fames ac turpis egestas.
-            Mauris pellentesque nibh id erat pulvinar vestibulum. Duis id nisl
-            mi. Praesent eu justo dolor. Fusce congue lectus et turpis varius
-            accumsan. Vestibulum porta libero sit amet magna tincidunt
-            vestibulum. Sed quis lobortis nulla. Phasellus in mattis tellus.
-            Vestibulum dolor mi, vehicula sed iaculis ac, blandit sit amet
-            nulla. Sed sollicitudin eget ex sit amet dapibus.
-          </p>
-          <p>
-            Phasellus sem risus, fermentum nec metus at, sagittis maximus nulla.
-            Curabitur mattis metus eu sapien tincidunt dictum. Nullam tempor
-            mattis dui, vel pellentesque tortor. Ut tellus velit, placerat et
-            magna nec, pellentesque maximus lacus. Vestibulum ac leo sed sem
-            ornare laoreet eget vel eros. Pellentesque habitant morbi tristique
-            senectus et netus et malesuada fames ac turpis egestas. Curabitur
-            dui mi, commodo et consequat a, congue at mauris. Orci varius
-            natoque penatibus et magnis dis parturient montes, nascetur
-            ridiculus mus. In molestie ipsum ante, id porttitor lacus lobortis
-            eu. Mauris finibus ligula ipsum, sed lacinia mi posuere in. Sed
-            placerat dapibus odio, feugiat mollis orci pulvinar mollis. Integer
-            malesuada condimentum dignissim. Duis dictum risus felis, at
-            ullamcorper ipsum pellentesque eget. Nullam sit amet dolor
-            porttitor, aliquet ante eget, sagittis felis. Phasellus aliquam dui
-            quam, quis dapibus lacus laoreet id.
-          </p>
-          <p>
-            Nullam pretium fringilla nunc sit amet mattis. Aliquam id libero
-            quam. Sed mollis nisl sed diam dictum, vitae accumsan nulla
-            tincidunt. Sed felis elit, pellentesque id mattis vitae, blandit sit
-            amet massa. Vestibulum accumsan neque mi, sit amet dapibus nisl
-            condimentum a. Proin ut maximus magna. Sed hendrerit neque quis
-            augue ornare tincidunt. Proin at augue eu mi lacinia sollicitudin
-            non a mi. In ac ornare risus. Proin ac eleifend sem. Nulla quis
-            finibus lacus. Suspendisse iaculis semper consectetur. Curabitur
-            volutpat scelerisque arcu, at posuere nibh imperdiet ac.
-          </p>
-          <hr></hr>
-          <p>
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas ut
-            mattis ligula. Maecenas congue, ipsum nec congue scelerisque, risus
-            arcu placerat ligula, at fringilla massa justo ut erat. Mauris sed
-            tellus est. Aenean fringilla, lectus vitae vehicula sollicitudin,
-            mauris ex tempus dui, in consectetur nisl quam ac felis. Donec eu
-            laoreet velit, sed fermentum lectus. Pellentesque habitant morbi
-            tristique senectus et netus et malesuada fames ac turpis egestas.
-            Mauris pellentesque nibh id erat pulvinar vestibulum. Duis id nisl
-            mi. Praesent eu justo dolor. Fusce congue lectus et turpis varius
-            accumsan. Vestibulum porta libero sit amet magna tincidunt
-            vestibulum. Sed quis lobortis nulla. Phasellus in mattis tellus.
-            Vestibulum dolor mi, vehicula sed iaculis ac, blandit sit amet
-            nulla. Sed sollicitudin eget ex sit amet dapibus.
-          </p>
-          <p>
-            Phasellus sem risus, fermentum nec metus at, sagittis maximus nulla.
-            Curabitur mattis metus eu sapien tincidunt dictum. Nullam tempor
-            mattis dui, vel pellentesque tortor. Ut tellus velit, placerat et
-            magna nec, pellentesque maximus lacus. Vestibulum ac leo sed sem
-            ornare laoreet eget vel eros. Pellentesque habitant morbi tristique
-            senectus et netus et malesuada fames ac turpis egestas. Curabitur
-            dui mi, commodo et consequat a, congue at mauris. Orci varius
-            natoque penatibus et magnis dis parturient montes, nascetur
-            ridiculus mus. In molestie ipsum ante, id porttitor lacus lobortis
-            eu. Mauris finibus ligula ipsum, sed lacinia mi posuere in. Sed
-            placerat dapibus odio, feugiat mollis orci pulvinar mollis. Integer
-            malesuada condimentum dignissim. Duis dictum risus felis, at
-            ullamcorper ipsum pellentesque eget. Nullam sit amet dolor
-            porttitor, aliquet ante eget, sagittis felis. Phasellus aliquam dui
-            quam, quis dapibus lacus laoreet id.
-          </p>
-          <p>
-            Nullam pretium fringilla nunc sit amet mattis. Aliquam id libero
-            quam. Sed mollis nisl sed diam dictum, vitae accumsan nulla
-            tincidunt. Sed felis elit, pellentesque id mattis vitae, blandit sit
-            amet massa. Vestibulum accumsan neque mi, sit amet dapibus nisl
-            condimentum a. Proin ut maximus magna. Sed hendrerit neque quis
-            augue ornare tincidunt. Proin at augue eu mi lacinia sollicitudin
-            non a mi. In ac ornare risus. Proin ac eleifend sem. Nulla quis
-            finibus lacus. Suspendisse iaculis semper consectetur. Curabitur
-            volutpat scelerisque arcu, at posuere nibh imperdiet ac.
-          </p>
-        </div>
-      </div>
-      <div className="sidebar">
-        <div className="about">
-          <h2>About</h2>
-          <p>
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas ut
-            mattis ligula. Maecenas congue, ipsum nec congue scelerisque, risus
-            arcu placerat ligula, at fringilla massa justo ut erat. Mauris sed
-            tellus est. Aenean fringilla, lectus vitae vehicula sollicitud
-          </p>
-        </div>
-        <div className="trends">
-          <h2>Trends</h2>
-          <ul className="trends-list">
-            <li className="trend">
-              <img src="/vercel.svg"></img>
-              <div>
-                <h3>Exemple Post</h3>
-                <p>January 1, 2024</p>
-              </div>
-            </li>
-          </ul>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-export default MainHub
