--- containers/MoustRecent/index.tsx
@@ -1,12 +1,0 @@
-import Card from "@/components/Areas/Card";
-
-const MoustRecent = () => {
-    return (
-        <div className="card-area">
-            <Card></Card>
-            <Card></Card>
-        </div>
-    );
-}
-
-export default MoustRecent;
